Do not modify score at all if no points were gained
This returns early if no points were gained, to avoid treating the
current score like a new score that has been arrived at.

This fixes a problem whereby a player with points that are a multiple
of 50 would have their points halve if they gained no new points in
a round.

Fixes #7